Initial CHERI support

+++ libunwind/src/UnwindLevel1-gcc-ext.c
@@ -94,7 +94,7 @@ _LIBUNWIND_EXPORT void *_Unwind_FindEnclosingFunction(void *pc) {
   unw_init_local(&cursor, &uc);
   unw_set_reg(&cursor, UNW_REG_IP, (unw_word_t)(long) pc);
   if (unw_get_proc_info(&cursor, &info) == UNW_ESUCCESS)
-    return (void *)(long) info.start_ip;
+    return (void *) info.start_ip;
   else
     return NULL;
 }
@@ -196,8 +196,8 @@ _LIBUNWIND_EXPORT const void *_Unwind_Find_FDE(const void *pc,
   bases->dbase = 0; // dbase not used on Mac OS X
   bases->func = (uintptr_t)info.start_ip;
   _LIBUNWIND_TRACE_API("_Unwind_Find_FDE(pc=%p) => %p", pc,
-                  (void *)(long) info.unwind_info);
-  return (void *)(long) info.unwind_info;
+                  (void *) info.unwind_info);
+  return (void *) info.unwind_info;
 }
 
 /// Returns the CFA (call frame area, or stack pointer at start of function)
